+++ public/dev/js/science_portal.js
@@ -41,7 +41,6 @@
     var portalSessions = new cadc.web.science.portal.session.PortalSession(inputs)
     var portalForm = new cadc.web.science.portal.form.PortalForm(inputs)
 
-
     var _selfPortalApp = this
 
     this.baseURL = inputs.baseURL
@@ -55,7 +54,7 @@
       this.isDev = inputs.isDev
     }
 
-    // TODO: this is for dev or situations where a registry app
+    // NOTE: this is for dev or situations where a registry app
     // might not be available
     var URLOverrides = inputs.URLOverrides
 
@@ -63,9 +62,6 @@
     var _state = {
       currentSessionData: {}
     }
-
-    // Complete list of form fields available
-    //var _launchFormFields = ["name", "type", "image", "memory", "cores"]
     var _curSessionType
 
     // ------------ Page load functions ------------
@@ -78,6 +74,7 @@
       portalCore.setHeaderURLs()
       portalForm.setContentBase(contentBase)
       portalForm.loadSessionTypeMap()
+      portalCore.setPageState("all", "primary", true, "")
     }
 
     function continueInit() {
@@ -89,8 +86,8 @@
       portalCore.initConfirmModal(_reactApp, handleConfirmedDelete, handleCancelDelete)
 
       if (typeof inputs.bannerText !== "undefined") {
-      _reactApp.setBanner(inputs.bannerText)
-    }
+        _reactApp.setBanner(inputs.bannerText)
+      }
 
       // Nothing happens if user is not authenticated, so no other page
       // load information is done until this call comes back (see onAuthenticated event below)
@@ -101,6 +98,7 @@
     function attachListeners() {
       // Button/page click listeners
       $(".sp-session-reload").click(checkForSessions)
+      $(".sp-e-usage-reload").click(handlePlatformUsageLoad)
 
       // Data Flow/javascript object listeners
       // portalCore listeners
@@ -120,13 +118,16 @@
         // Get portalForm to start collecting form data
         portalForm.getFormData()
 
+        // Get platform Usage information
+        portalSessions.loadPlatformUsage()
+
         // Start loading session lists
         checkForSessions()
       })
 
       portalCore.subscribe(portalCore, cadc.web.science.portal.core.events.onServiceURLFail, function (e, data){
         // Unable to contact registry to get sessions web service URL
-        portalCore.setPageState("danger", false)
+        portalCore.setPageState(portalCore.pageSections.all,"danger", false)
 
         portalCore.setModal(_reactApp, "Portal Unavailable",
           "Unable to establish communication with Skaha web service. "
@@ -140,15 +141,15 @@
         // Build session list on top of page
         var filteredList = portalSessions.getFilteredSessionList()
         populateSessionList(filteredList)
-        portalCore.setPageState("success", false)
-        portalCore.hideModal(_reactApp)
+        portalCore.setPageState(portalCore.pageSections.sessionList, "success", false)
+        portalCore.hideModal()
 
         if (( _selfPortalApp.isPolling === false)
           && (portalSessions.isAllSessionsStable() === false) ){
 
           // Flag polling is occurring so only one instance is running at a time.
           // Any changes in session list will be picked up by the single polling instance
-          _selfPortalApp.isPolling == true
+          _selfPortalApp.isPolling = true
 
           // If everything is stable, stop. If no, kick off polling
           portalSessions.pollSessionList(8000)
@@ -177,9 +178,13 @@
       })
 
       portalCore.subscribe(portalSessions, cadc.web.science.portal.session.events.onLoadSessionListError, handleServiceError)
-      portalCore.subscribe(_selfPortalApp, cadc.web.science.portal.events.onSessionRequestOK, refreshSessionForm)
       portalCore.subscribe(portalSessions, cadc.web.science.portal.session.events.onSessionActionDone, refreshSessionForm)
       portalCore.subscribe(portalSessions, cadc.web.science.portal.session.events.onSessionActionError, handleSessionActionError)
+      portalCore.subscribe(portalSessions, cadc.web.science.portal.session.events.onLoadPlatformUsageDone, populatePlatformUsage)
+      portalCore.subscribe(portalSessions, cadc.web.science.portal.session.events.onLoadPlatformUsageError, handleSessionActionError)
+
+      // Listeners for this class (science_portal.js)
+      portalCore.subscribe(_selfPortalApp, cadc.web.science.portal.events.onSessionRequestOK, refreshSessionForm)
 
       // Portal Form listeners
       portalCore.subscribe(portalForm, cadc.web.science.portal.form.events.onLoadFormDataDone, initForm)
@@ -189,6 +194,8 @@
 
 
     } // end attachListeners()
+
+
 
     function handleServiceError(e, request) {
 
@@ -209,8 +216,16 @@
       portalCore.setModal(_reactApp, msgHeader, msgBody, false, true, true)
     }
 
+    function handlePlatformUsageLoad() {
+      // Leave the list as is, and update the progress bar to show something is happening.
+      // portalCore.setModal(_reactApp,"Statistics Check", "Fetching platform usage", true, false, false)
+      portalCore.clearAjaxAlert(portalCore.pageSections.usage)
+      portalCore.setPageState(portalCore.pageSections.usage, "primary", true, "")
+      portalSessions.loadPlatformUsage(_selfPortalApp.handlePlatformUsage)
+    }
+
     function handleSessionActionError(e, request) {
-      portalCore.setSessionActionAjaxFail(request)
+      portalCore.setAjaxFail(portalCore.pageSections.sessionList, request)
     }
 
     function refreshSessionForm() {
@@ -220,6 +235,16 @@
     }
 
     // ------------ Data display functions
+
+    function populatePlatformUsage() {
+      // reset page state
+      portalCore.setPageState(portalCore.pageSections.usage, "success", false)
+      portalCore.hideModal()
+      var platformUsage = portalSessions.getPlatformUsage()
+
+      // pass platformUsage info to the react App for rendering
+      _reactApp.updatePlatformUsage(platformUsage)
+    }
 
     function populateSessionList(sessionData) {
       var $sessionListDiv = $("#sp_session_list")
@@ -255,10 +280,6 @@
               iconLabel = mapEntry.portal_text
             }
           }
-
-          // TODO: fill in actual RAM & cores values from /session endpoint
-          // once that update is available. Jan 2023: they are not displayed
-          // on the SessionItem cards currently.
 
           var nextSessionItem = {
             "id" : this.id,
@@ -297,7 +318,8 @@
 
     function checkForSessions() {
       portalCore.setModal(_reactApp,"Session Check", "Fetching session list", true, false, false)
-      portalCore.setPageState("success", true)
+      portalCore.clearAjaxAlert(portalCore.pageSections.sessionList)
+      portalCore.setPageState(portalCore.pageSections.sessionList, "primary", true)
       portalSessions.loadSessionList()
     }
 
@@ -312,7 +334,7 @@
       event.preventDefault()
       // Pull data-* information from anchor element
       var sessionData = $(event.currentTarget).data()
-      portalCore.setPageState("success", false)
+      portalCore.setPageState(portalCore.pageSections.all,"success", false)
       window.open(sessionData.connecturl, "_blank")
     }
 
@@ -328,7 +350,7 @@
      */
     function handleDeleteSession(event) {
       event.preventDefault()
-      portalCore.clearAjaxAlert()
+      portalCore.clearAjaxAlert(portalCore.pageSections.sessionList)
       var sessionData = event.currentTarget.dataset
       _state.currentSessionData = sessionData
       portalCore.setConfirmModal(_reactApp, handleConfirmedDelete, sessionData)
@@ -355,16 +377,13 @@
      * Triggered from the renew icon on a session card
      */
     function handleRenewSession(event) {
-      portalCore.clearAjaxAlert()
+      portalCore.clearAjaxAlert(portalCore.pageSections.sessionList)
       var sessionData = event.currentTarget.dataset
       portalCore.setModal(_reactApp, "Renew Session Request", "Extending session time", true, false, false)
       portalSessions.renewSession(sessionData.id)
     }
 
-
-
     // ------------ HTTP/Ajax functions & event handlers ------------
-    // ---------------- POST ------------------
 
     /**
      * Submit button function
@@ -373,7 +392,7 @@
     function handleSessionRequest(event) {
       // Stop normal form submit
       event.preventDefault()
-      portalCore.clearAjaxAlert(_reactApp)
+      portalCore.clearAjaxAlert(portalCore.pageSections.form)
 
       var _prunedFormData = new FormData();
 
@@ -383,15 +402,16 @@
         console.log(currentValue.name + ": " + currentValue.value)
       }
 
+      portalCore.setPageState(portalCore.pageSections.form, "primary", true, '')
       portalCore.setModal(_reactApp, "Requesting Session", "Requesting new session", true, false, false)
       Promise.resolve(postSessionRequestAjax(portalCore.sessionServiceURLs.session, _prunedFormData))
         .then(function(sessionInfo) {
-          portalCore.setPageState("success", false)
+          portalCore.setPageState(portalCore.pageSections.form,"success", false)
           portalCore.hideModal(_reactApp)
           portalCore.trigger(_selfPortalApp, cadc.web.science.portal.events.onSessionRequestOK, sessionInfo)
         })
         .catch(function(request) {
-          portalCore.handleAjaxError(request, _reactApp)
+          portalCore.handleAjaxError(portalCore.pageSections.form, request)
         })
     }
 
@@ -429,8 +449,6 @@
     // This can only happen after the portalForm has grabbed all the data
     // triggered when onLoadFormDataDone is issued
     function initForm() {
-      //_curSessionType = portalForm.getSessionTypeDefault()
-      //setLaunchForm(_curSessionType)
       resetLaunchForm()
     }
 
@@ -452,7 +470,7 @@
     }
 
     function resetLaunchForm() {
-      portalCore.clearAjaxAlert(_reactApp)
+      portalCore.clearAjaxAlert()
       _curSessionType = portalForm.getSessionTypeDefault()
       setLaunchForm(_curSessionType)
     }